+++ src/components/Spinner.tsx
@@ -1,9 +1,9 @@
-export default function Spinner() {
+export default function Spinner({ size = 8 }: { size?: number }) {
   return (
     <div role="status">
       <svg
         aria-hidden="true"
-        className="w-8 h-8 text-gray-200 animate-spin dark:text-gray-600 fill-primary"
+        className={`w-${size} h-${size} text-gray-200 animate-spin dark:text-gray-600 fill-primary`}
         viewBox="0 0 100 101"
         fill="none"
         xmlns="http://www.w3.org/2000/svg"
